requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Document parsing
pymupdf>=1.23.0
python-docx>=1.1.0

# Vector store (using numpy for simple in-memory store)
numpy>=1.26.0

# Embeddings
openai>=1.12.0
sentence-transformers>=2.3.0

# Configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
